third_party/WebKit/LayoutTests/http/tests/media/media-source/webkitmediasource-duration-boundaryconditions-expected.txt

CONSOLE WARNING: 'WebKitMediaSource' is deprecated. Please use 'MediaSource' instead.
Tests boundary values for duration attribute on MediaSource object.


running testSetMaxInt
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
EXPECTED (mediaSource.duration == 'testDurationValue') OK
onDurationChange : video.duration is 2147483647.000
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(durationchange)

running testSetMinInt
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
EXPECTED (mediaSource.duration == 'testDurationValue') OK
onDurationChange : video.duration is 1.000
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(durationchange)

running testSetMaxValue
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
EXPECTED (mediaSource.duration == 'testDurationValue') OK
onDurationChange : video.duration is 1.7976931348623157e+308
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(durationchange)

running testSetMaxValueMinusOne
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
EXPECTED (mediaSource.duration == 'testDurationValue') OK
onDurationChange : video.duration is 1.7976931348623157e+308
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(durationchange)

running testSetMinValue
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
EXPECTED (mediaSource.duration == 'testDurationValue') OK
onDurationChange : video.duration is 0.000
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(durationchange)

running testSetMinValueMinusOne
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
PASS: Media Source set duration. Got expected exception InvalidAccessError: Failed to set the 'duration' property on 'WebKitMediaSource': The duration provided (-1) is less than the minimum bound (0).

running testSetPositiveInfinity
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
EXPECTED (mediaSource.duration == 'testDurationValue') OK
onDurationChange : video.duration is Infinity
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(durationchange)

running testSetNegativeInfinity
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
PASS: Media Source set duration. Got expected exception InvalidAccessError: Failed to set the 'duration' property on 'WebKitMediaSource': The duration provided (-Infinity) is less than the minimum bound (0).

running testSetLowestValue
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
PASS: Media Source set duration. Got expected exception InvalidAccessError: Failed to set the 'duration' property on 'WebKitMediaSource': The duration provided (-1.797693e+308) is less than the minimum bound (0).

running testSetNegativeDouble
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
PASS: Media Source set duration. Got expected exception InvalidAccessError: Failed to set the 'duration' property on 'WebKitMediaSource': The duration provided (-101.9) is less than the minimum bound (0).

running testSetPositiveDouble
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
EXPECTED (mediaSource.duration == 'testDurationValue') OK
onDurationChange : video.duration is 101.900
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(durationchange)

running testSetZero
EVENT(webkitsourceopen)
onDurationChange : video.duration is 6.042
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(loadedmetadata)
EXPECTED (mediaSource.duration == 'testDurationValue') OK
onDurationChange : video.duration is 0.000
EXPECTED (mediaSource.duration == 'video.duration') OK
EVENT(durationchange)
END OF TEST
